+++ kompendium-swagger-ui/src/main/kotlin/io/bkbn/kompendium/swagger/SwaggerWebJarUtils.kt
@@ -1,6 +1,8 @@
 package io.bkbn.kompendium.swagger
 
 import io.ktor.features.NotFoundException
+import io.ktor.http.ContentType
+import io.ktor.http.defaultForFilePath
 import io.ktor.http.content.ByteArrayContent
 import java.net.URL
 import org.webjars.WebJarAssetLocator
@@ -10,10 +12,14 @@ internal fun WebJarAssetLocator.getSwaggerResource(path: String): URL =
     ?: throw NotFoundException("Resource not found: $path")
 
 internal fun WebJarAssetLocator.getSwaggerResourceContent(path: String): ByteArrayContent =
-  ByteArrayContent(getSwaggerResource(path).readBytes())
+    ByteArrayContent(
+        contentType = ContentType.defaultForFilePath(path),
+        bytes = getSwaggerResource(path).readBytes()
+    )
 
 internal fun WebJarAssetLocator.getSwaggerInitializerContent(jsConfig: JsConfig): ByteArrayContent = ByteArrayContent(
-    getSwaggerResource(path = "swagger-initializer.js").readText()
+    contentType = ContentType.Application.JavaScript,
+    bytes = getSwaggerResource(path = "swagger-initializer.js").readText()
       .replaceFirst("url: \"https://petstore.swagger.io/v2/swagger.json\",", "urls: ${jsConfig.getSpecUrlsProps()},")
       .replaceFirst("deepLinking: true", jsConfig.toJsProps())
       .let { content ->